refactor #3 - 예산 조회 및 추천 기능 리팩토링 - BigDecimal에서 Integer 사용하는 것으로 변경 - Budget 엔티티에 rate 컬럼 추가 - 카테고리 평균 비율 계산시 BudgetRepository 이용

+++ src/main/java/com/example/budget/dto/req/BudgetCreateReqDto.java
@@ -3,7 +3,6 @@ package com.example.budget.dto.req;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import jakarta.validation.constraints.NotEmpty;
-import java.math.BigDecimal;
 import java.util.List;
 import lombok.Builder;
 
@@ -11,14 +10,14 @@ import lombok.Builder;
 public record BudgetCreateReqDto(
     @NotEmpty
     List<BudgetDto> budgets,
-    BigDecimal totalAmount
+    Integer totalAmount
 ) {
 
   @Builder
   @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
   public record BudgetDto(
       Long categoryId,
-      BigDecimal amount
+      Integer amount
   ) {
 
   }
